Extract SSRF-hardened fetch path from WebFetch into shared SafeFetch
Behavior-preserving refactor. All URL validation, hostname/IP safety
checks, the GlobalOnlyResolver session factory, content-type probing,
redirect refusal, size caps, and the streamed PDF download move from
private methods on the WebFetch coded tool to the shared utility
neuro_san_studio/coded_tools/utils/safe_fetch.py (SafeFetch), a sibling
of PdfUtils. WebFetch keeps only tool-level policy (character cap,
supported content types, arg validation) and delegates the rest.

This makes the hardened fetch path reusable by other coded tools that
retrieve remote content; webpage RAG (#1300) and PDF RAG (#1299) build
on it next.

The existing fetch-machinery unit tests move to
tests/neuro_san_studio/coded_tools/utils/safe_fetch/ with assertions
unchanged, so they double as proof the extraction preserves behavior.

Part of #1300.
