+++ tests/neuro_san_studio/coded_tools/utils/safe_fetch/helpers.py
