+++ tests/neuro_san_studio/coded_tools/web_fetch/test_async_invoke.py
@@ -19,6 +19,7 @@ from unittest import TestCase
 from unittest.mock import AsyncMock
 from unittest.mock import patch
 
+from neuro_san_studio.coded_tools.utils.safe_fetch import SafeFetch
 from neuro_san_studio.coded_tools.web_fetch import WebFetch
 
 
@@ -37,8 +38,8 @@ class TestAsyncInvoke(TestCase):
     def test_html_fetch_returns_correct_keys(self):
         """Tests that fetching an HTML page returns a result with url, content, and retrieved_at keys."""
         with (
-            patch.object(self.tool, "_get_content_type", new=AsyncMock(return_value=("text/html", None))),
-            patch.object(self.tool, "_fetch_text", new=AsyncMock(return_value="Hello world")),
+            patch.object(SafeFetch, "get_content_type", new=AsyncMock(return_value=("text/html", None))),
+            patch.object(SafeFetch, "fetch_text", new=AsyncMock(return_value="Hello world")),
         ):
             result = asyncio.run(self.tool.async_invoke({"url": "http://example.com"}, self.sly_data))
 
@@ -47,12 +48,12 @@ class TestAsyncInvoke(TestCase):
         self.assertIn("retrieved_at", result)
 
     def test_405_prefetched_body_skips_fetch_text(self):
-        """Tests that a prefetched body from the 405 GET fallback is used directly without calling _fetch_text."""
+        """Tests that a prefetched body from the 405 GET fallback is used directly without calling fetch_text."""
         with (
             patch.object(
-                self.tool, "_get_content_type", new=AsyncMock(return_value=("text/html", "<p>prefetched</p>"))
+                SafeFetch, "get_content_type", new=AsyncMock(return_value=("text/html", "<p>prefetched</p>"))
             ),
-            patch.object(self.tool, "_fetch_text", new=AsyncMock(return_value="should not be called")) as mock_text,
+            patch.object(SafeFetch, "fetch_text", new=AsyncMock(return_value="should not be called")) as mock_text,
         ):
             result = asyncio.run(self.tool.async_invoke({"url": "http://example.com"}, self.sly_data))
 
@@ -60,11 +61,11 @@ class TestAsyncInvoke(TestCase):
         self.assertIn("prefetched", result["content"])
 
     def test_pdf_by_content_type_calls_fetch_pdf(self):
-        """Tests that an application/pdf content type routes to _fetch_pdf and not _fetch_text."""
+        """Tests that an application/pdf content type routes to SafeFetch.fetch_pdf_text and not fetch_text."""
         with (
-            patch.object(self.tool, "_get_content_type", new=AsyncMock(return_value=("application/pdf", None))),
-            patch.object(self.tool, "_fetch_pdf", new=AsyncMock(return_value="PDF content")) as mock_pdf,
-            patch.object(self.tool, "_fetch_text", new=AsyncMock(return_value="should not be called")) as mock_text,
+            patch.object(SafeFetch, "get_content_type", new=AsyncMock(return_value=("application/pdf", None))),
+            patch.object(SafeFetch, "fetch_pdf_text", new=AsyncMock(return_value="PDF content")) as mock_pdf,
+            patch.object(SafeFetch, "fetch_text", new=AsyncMock(return_value="should not be called")) as mock_text,
         ):
             result = asyncio.run(self.tool.async_invoke({"url": "http://example.com/file"}, self.sly_data))
 
@@ -73,12 +74,12 @@ class TestAsyncInvoke(TestCase):
         self.assertEqual(result["content"], "PDF content")
 
     def test_pdf_by_url_extension_calls_fetch_pdf(self):
-        """Tests that a .pdf URL extension routes to _fetch_pdf regardless of content type."""
+        """Tests that a .pdf URL extension routes to fetch_pdf_text regardless of content type."""
         with (
             patch.object(
-                self.tool, "_get_content_type", new=AsyncMock(return_value=("application/octet-stream", None))
+                SafeFetch, "get_content_type", new=AsyncMock(return_value=("application/octet-stream", None))
             ),
-            patch.object(self.tool, "_fetch_pdf", new=AsyncMock(return_value="PDF content")) as mock_pdf,
+            patch.object(SafeFetch, "fetch_pdf_text", new=AsyncMock(return_value="PDF content")) as mock_pdf,
         ):
             asyncio.run(self.tool.async_invoke({"url": "http://example.com/report.pdf"}, self.sly_data))
 
@@ -86,7 +87,7 @@ class TestAsyncInvoke(TestCase):
 
     def test_unsupported_content_type_raises(self):
         """Tests that an unsupported content type raises ValueError with unsupported_content_type."""
-        with patch.object(self.tool, "_get_content_type", new=AsyncMock(return_value=("image/png", None))):
+        with patch.object(SafeFetch, "get_content_type", new=AsyncMock(return_value=("image/png", None))):
             with self.assertRaises(ValueError) as ctx:
                 asyncio.run(self.tool.async_invoke({"url": "http://example.com/img.png"}, self.sly_data))
         self.assertIn("unsupported_content_type", str(ctx.exception))
@@ -95,8 +96,8 @@ class TestAsyncInvoke(TestCase):
         """Tests that fetched content is truncated to the specified max_content_chars limit."""
         long_text = "x" * 1000
         with (
-            patch.object(self.tool, "_get_content_type", new=AsyncMock(return_value=("text/plain", None))),
-            patch.object(self.tool, "_fetch_text", new=AsyncMock(return_value=long_text)),
+            patch.object(SafeFetch, "get_content_type", new=AsyncMock(return_value=("text/plain", None))),
+            patch.object(SafeFetch, "fetch_text", new=AsyncMock(return_value=long_text)),
         ):
             result = asyncio.run(
                 self.tool.async_invoke({"url": "http://example.com", "max_content_chars": 100}, self.sly_data)
